refactor(vital_records): move email task to subpackage
